refactor(examples): remove directory fs these are the last internal usages of the package that's about to be deprecated
