+++ examples/file-server/package.json
@@ -8,7 +8,6 @@
     "test:unit": "mocha \"dist/test/**/*.unit.js\""
   },
   "dependencies": {
-    "@file-services/directory": "^7.0.1",
     "@file-services/node": "^7.0.1"
   },
   "license": "UNLICENSED",
